infob-auomation-v2/requirements.txt
requests
google-auth
google-cloud-secret-manager